change ttc value to integer

+++ gridcapa-swe-runner-app/src/main/java/com/farao_community/farao/swe/runner/app/ttc_doc/TtcDocument.java
@@ -68,7 +68,7 @@ public class TtcDocument {
 
     private void addValueToResultMap(DichotomyDirection direction, DichotomyResult<RaoResponse> result) {
         if (result.hasValidStep()) {
-            mapWithValues.put(direction, String.valueOf(result.getHighestValidStepValue()));
+            mapWithValues.put(direction, String.valueOf((int) result.getHighestValidStepValue()));
         } else {
             mapWithValues.put(direction, "");
         }
